backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
pytz==2024.1
email-validator==2.1.0.post1 